* master: (27 commits)
  Add configure option to disable man pages
  Remove debug log
  Improve logging related to hashing of the CWD
  Improve installation instructions
  Update NEWS
  Avoid TOCTOU issue when deciding if config is valid
  Fix code style
  hashutil: fix memory-leak/double free
  Fix some casting warnings
  Update NEWS
  Add the sanitize blacklist contents to the hash
  Update NEWS
  Simplify logic
  Print the time when stats last updated
  Use double when calculating cache thresholds
  Add a 32-bit build target, using multilib (-m32)
  Convert float config to double, add rounding
  Upgrade clang to 5.0, for the Travis docker
  Make sure to export ASAN_OPTIONS properly
  Add small helper to run all Travis tests
  ...

+++ src/util.c
@@ -1053,7 +1053,7 @@ parse_size_with_suffix(const char *str, uint64_t *size)
 		// Default suffix: G.
 		x *= 1000 * 1000 * 1000;
 	}
-	*size = x;
+	*size = (uint64_t)x;
 	return true;
 }
